refactor metrics grouping
