+++ server/metrics.js
@@ -1,11 +1,11 @@
 export function aggregateByMonth(vacancies) {
-  const groups = {};
+  const groups = new Map();
   for (const v of vacancies) {
     const month = v.date.slice(0, 7);
-    groups[month] = groups[month] || [];
-    groups[month].push(v);
+    if (!groups.has(month)) groups.set(month, []);
+    groups.get(month).push(v);
   }
-  return Object.entries(groups)
+  return Array.from(groups.entries())
     .sort(([a], [b]) => a.localeCompare(b))
     .map(([period, items]) => {
       const posted = items.length;
@@ -18,13 +18,3 @@ export function aggregateByMonth(vacancies) {
       return { period, posted, awarded, cancelled, cancellationRate, overtime };
     });
 }
-
-export const sampleVacancies = [
-  { date: '2024-01-01', status: 'awarded', hours: 10 },
-  { date: '2024-01-05', status: 'cancelled', hours: 8 },
-  { date: '2024-01-07', status: 'posted', hours: 8 },
-  { date: '2024-02-02', status: 'awarded', hours: 9 },
-  { date: '2024-02-04', status: 'awarded', hours: 8 },
-  { date: '2024-02-08', status: 'cancelled', hours: 8 },
-  { date: '2024-03-03', status: 'posted', hours: 8 }
-];
