+++ server/sampleData/vacancies.ts
@@ -1,0 +1,9 @@
+export const sampleVacancies = [
+  { date: '2024-01-01', status: 'awarded', hours: 10 },
+  { date: '2024-01-05', status: 'cancelled', hours: 8 },
+  { date: '2024-01-07', status: 'posted', hours: 8 },
+  { date: '2024-02-02', status: 'awarded', hours: 9 },
+  { date: '2024-02-04', status: 'awarded', hours: 8 },
+  { date: '2024-02-08', status: 'cancelled', hours: 8 },
+  { date: '2024-03-03', status: 'posted', hours: 8 }
+];
